Moved SkillHttpClient to correct library with corresponding tests. Fix SkillDialog.
